+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [5.1.0]
+- Added new colors from Design System.
+- Components app: Added better download message.
+- Components app: Added search for resources.
+
 ## [5.0.0]
 - Added build script for delivering Components app to AppCenter
 - Added in app API for checking version of AppCenter